big fix

+++ Student/del.php
@@ -1,11 +1,13 @@
 <?php
     include "../conn.php";
 
-    $code=$_GET['id'];
-    $order=$_GET['ord'];
+    $code = g('id');
+    $order = (int) g('ord');
 
-    $qry = "DELETE FROM `orders` WHERE `Order_ID` = $order";
-    $result = mysqli_query($conn, $qry);
+    if ($order > 0) {
+        $qry = "DELETE FROM `orders` WHERE `Order_ID` = $order";
+        mysqli_query($conn, $qry);
+    }
     header("location: history.php?code=$code&type=student_rgo");
     exit;
     ?>
